refactor: implement a decomposition of the `welcome page`
