+++ src/components/icons/RSSchoolIcon.tsx
@@ -1,0 +1,6 @@
+import RsSchool from '@/assets/welcomePage/rs_school_js.svg?react';
+import { createSvgIcon } from '@mui/material';
+
+const RSSchoolIcon = createSvgIcon(<RsSchool />, 'RSSchool');
+
+export default RSSchoolIcon;
