+++ src/localization/locales/en.json
@@ -10,6 +10,7 @@
   "welcomeNotAuth": "To try it you must be authenticated",
   "welcomeAuth": "Just try it",
   "ourSchool": "Our School",
+  "schoolBio": "RS School is free-of-charge and community-based education program conducted by The Rolling Scopes developer community since 2013.\nEveryone can study at RS School, regardless of age, professional employment, or place of residence. \nThe mentors and trainers of our school are front-end and javascript developers from different companies and countries.",
   "ourTeam": "Our Team",
   "position": "Position",
   "developer": "Development",
