+++ src/pages/welcome/components/styles/styles.ts
@@ -1,0 +1,13 @@
+import { Theme } from '@emotion/react';
+import { SxProps } from '@mui/material';
+
+export const greySection: SxProps<Theme> = {
+  bgcolor: '#e0dee3',
+  width: '100%',
+};
+
+export const flexRowWrap: SxProps<Theme> = {
+  display: 'flex',
+  flexDirection: 'row',
+  flexWrap: 'wrap',
+};
